// src/types.ts
export interface AccountResponse {
  ranking: number;
  bankImage: string;
  bankingProduct: string;
  validFrom: string;
  base: string;
  maxBonus: string;
  maxTotal: string;
  requirementForBonus: string;
  maxBalance: string;
  comments: string;
  effectiveAsOf: string;
  lastCheckedOrEdited: string;
  link: string;
  category: string;
}